requirements.txt
fastapi==0.104.1
pydantic==2.4.2
uvicorn==0.24.0
passlib==1.7.4
PyJWT==2.8.0
bcrypt==4.0.1
python-multipart==0.0.6